fix(MessageAnAgent Node): Version the node so the agent picker ships as v2 (no-changelog)
Switching agentId from resourceLocator to agentSelector is a breaking change for
the v1 node already accessible in cloud (per review). Convert MessageAnAgent to a
VersionedNodeType:

- v1 keeps the original resourceLocator picker + listAgents search method, so
  existing workflows (typeVersion 1) are unaffected.
- v2 uses the new agentSelector picker.
- defaultVersion 2, so newly added nodes get the picker.

execute(), structured-output parsing, and every property except agentId are
shared in shared.ts. Addresses yehorkardash's review comment.
